#139 Decomposes functions to calculate measures and move them to another file
